requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
lxml==4.9.2
python-dotenv==1.0.0
requests==2.31.0
tabulate==0.9.0
tqdm==4.65.0
urllib3==2.0.3
